Lets deploy

+++ src/main/client/src/edituser.jsx
@@ -1,8 +1,8 @@
 import {useEffect, useState} from "react";
 import {useNavigate} from "react-router-dom";
 
-export function EditUser({activeUserId}) {
-    console.log(activeUserId)
+export function EditUser({activeUserId,setUpdate}) {
+
     const [loading, setLoading] = useState(true);
     const navigate = useNavigate();
     const [name, setName] = useState("");
@@ -24,38 +24,70 @@ export function EditUser({activeUserId}) {
 
             setLoading(false);
         }
-        test()
+        if (activeUserId>0){
+            test()
+        }
+        else{
+            //GET NEW USER
+            setLoading(false)
+        }
+
     }, []);
     if (loading){
         return  <h1>LOADING ..</h1>
     }
     async function handleOnSubmit(event) {
-
-
+        console.log("username"+username)
 
         event.preventDefault();
-
-
-        const tmp={
-            id_user: activeUserId,
-            name,
-            username,
-            tlf,
-            emails:emails
+        if (username.length < 2){
+            navigate("/newuser")
         }
-
-        const user = JSON.stringify(tmp)
-
-        const test= await fetch("/api/users", {
-            method: "PUT",
-
-            body: user,
-            headers: {
-                "Content-Type": "application/json"
+        if (activeUserId > 0){
+            const tmp={
+                id_user: activeUserId,
+                name,
+                username,
+                tlf,
+                emails:emails
             }
-        });
-        if (test.ok)
-            navigate("/user");
+
+            const user = JSON.stringify(tmp)
+
+            const test= await fetch("/api/users", {
+                method: "PUT",
+
+                body: user,
+                headers: {
+                    "Content-Type": "application/json"
+                }
+            });
+            if (test.ok)
+                navigate("/user");
+        }
+        else{
+            const tmp={
+                id_user: -1,
+                name,
+                username,
+                tlf,
+                emails:emails
+            }
+            const user = JSON.stringify(tmp)
+            const test= await fetch("/api/users", {
+                method: "POST",
+
+                body: user,
+                headers: {
+                    "Content-Type": "application/json"
+                }
+            });
+            setUpdate(true);
+            navigate("/");
+
+
+
+        }
 
     }
     async function handelRemoveEmail(e){
@@ -64,45 +96,52 @@ export function EditUser({activeUserId}) {
 
         const newEmails = emails.filter((email,index) =>  { return  index < emails.length-1})
         setEmails(newEmails);
-        console.log(emails)
-        await fetch("/api/users/emails/"+rmvEmail.id,{
-            method: 'DELETE',
-            headers: {
-                'Content-type': 'application/json'
-            }
-        })
+       if (activeUserId > 0){
+           await fetch("/api/users/emails/"+rmvEmail.id,{
+               method: 'DELETE',
+               headers: {
+                   'Content-type': 'application/json'
+               }
+           })
+       }
+       else{
+
+       }
+
     }
     async function handelAddMail(e) {
         e.preventDefault();
-        const index = emails.length-1;
-        const old = [...emails];
-        old[index].id
 
-        const params = {
-            id: activeUserId
-        };
 
-        const test = new URLSearchParams( {id: activeUserId})
-         await fetch("/api/users/emails",{
-            method: 'POST',
-            body: JSON.stringify({id_user:activeUserId}),
-            headers: {
-                'Content-type': 'application/json'
+        if(activeUserId > 0){
+
+            const res = await fetch("/api/users/emails");
+            const newid = await res.json();
+            console.log(newid);
+            const id = JSON.stringify(activeUserId);
+            await fetch("/api/users/emails/",{
+                method: 'POST',
+                body:id,
+                headers: {
+                    'Content-type': 'application/json'
+                }
+            })
+            const email ={
+                email:"exmple@exmple.no",
+                id:newid
             }
-        })
+            setEmails(prevState => [...prevState, email]);
 
-        const res = await fetch("/api/users/emails?"+new URLSearchParams( {id: activeUserId}));
-        const id = await res.json()
-
-        //console.log(id);
-
-        const email ={
-            email:"test",
-            id:id
+        }else{
+            const email ={
+                email:"test",
+                id:-1
+            }
+            setEmails(prevState => [...prevState, email]);
         }
 
 
-        setEmails(prevState => [...prevState, email]);
+
 
     }
 
@@ -116,37 +155,49 @@ export function EditUser({activeUserId}) {
 
     }
 
+    function handelNavigate(e) {
+        e.preventDefault()
+        if (activeUserId > 0)
+            navigate("/user")
+        else{
+            history.push('/');
+            window.location.reload(false);
+        }
+
+    }
+
     return (
         <div>
-            <h1>EDIT USER!</h1>
-
+           <center><div className={"editUserTilte"}>EDIT USER!</div></center>
+            <center>
             <form onSubmit={handleOnSubmit}>
-                <div>
-                    name:<input value={name} type="text" onChange={(e) => setName(e.target.value)}/>
+                <div className={"border"}>
+                    name:<input className="write-message" placeholder="Enter a Name" value={name} type="text" onChange={(e) => setName(e.target.value)}/>
                 </div>
-                <div>
-                    username:<input value={username} type="text" onChange={(e) => setUsername(e.target.value)}/>
+                <div className={"border"}>
+                    username:<input className="write-message" placeholder="Enter a Username" value={username} type="text" onChange={(e) => setUsername(e.target.value)}/>
                 </div>
-                <div>
-                    tlf:<input value={tlf} type="text" onChange={(e) => setTlf(e.target.value)}/>
+                <div className={"border"}>
+                    tlf:<input className="write-message" placeholder="Enter a phone nr" value={tlf} type="text" onChange={(e) => setTlf(e.target.value)}/>
                 </div>
                 <br/>
 
                 <div>
                     {emails.map((email,index)=>{
 
-                        return <div>
-                            email {index+1} :<input  value={email.email} type="text"
+                        return <div className={"border"}>
+                            email {index+1} :<input  type="email" className="write-message" placeholder="Enter a Email" value={email.email}
                                                     onChange={(e) => handelSetEmail(e,index)}/>
                         </div>
                     })}
-                    <button onClick={(e)=>handelAddMail(e)}>add mail</button><button onClick={(e)=>handelRemoveEmail(e)}>remove mail</button>
+                    <button className={"button"} onClick={(e)=>handelAddMail(e)}>add email</button><button className={"button"} onClick={(e)=>handelRemoveEmail(e)}>remove email</button>
                 </div>
 
-
-                <button>Update</button>
+                <button onClick={(e)=>handelNavigate(e)} className={"button"}>Back</button>
+                <button type="submit" className={"button"}>Update</button>
             </form>
 
+            </center>
         </div>
     )
 }
